pyproject.toml
[tool.poetry]
name = "test_backend"
version = "0.1.0"
description = ""
authors = ["werevolff <sirnikolasd@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.2"
Django = "^3.1.7"
django-environ = "0.4.5"
dj-database-url = "^0.5.0"
djangorestframework = "^3.12.4"
psycopg2 = "^2.8.6"
deepdiff = "^5.2.3"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "6.2.2"
black = {extras = ["d"], version = "^20.8b1"}
factory-boy = "^3.2.0"
pytest-factoryboy = "^2.1.0"
pytest-django = "^4.1.0"
pytest-drf = "^1.1.2"
PyHamcrest = "^2.0.2"
pytest-mock = "^3.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
